feat: Add support for waiting for non-cached values. (#160)
1. Allows for the common client to have customized data sources. This
was used for testing here, but could be used for more specialized
use-cases in the future.
2. Adds the ability for `start` and `identify` to wait for values from
LaunchDarkly instead of resolving with cached values when available.

With the previous wrapper SDK this behavior was controlled by the
underlying SDK and those would operate similar to the new, optional,
behavior.

For most uses cases it is both safer and higher performance to used
cached values when they are available instead of only using those cached
values as a fallback.

Ideally we would add new return values to start and identify. Next major
version we should consider:
`Cached`, `Timeout`. While retaining the current value for completing
via data from LD.