feat: Add experimental plugin support. (#225)
The flutter SDK is arranged into a common-client package that is
independent of the flutter package to allow for eventually supporting
non-flutter usages such as dart command line applications. This means,
similar to JS, that plugins are a concern of the leaf-node of the SDK
because plugins depend on types that are not common to the leaf nodes.

A plugin for the flutter client SDK is registered with an instance of
the flutter client SDK and not an instance of the common client SDK. (In
flutter the common client is actually composed into the leaf-node and
not a base class, so this difference is even more meaningful.)

The work-around to this is to make common types for all the meta-data,
and then a generic base class to be extended by the leaf-node
implementations. That extended version makes the client concrete instead
of generic. The base operations that need to be done on plugins, such as
getting a list of hooks from them, and registering them, are handled via
generic methods in the common client package.

```mermaid
  graph TD
      subgraph "common_client package"
          A[PluginBase&lt;TClient&gt;]
      end

      subgraph "flutter_client_sdk package"
          B[Plugin]
          B --> |extends| A
      end

      subgraph "User Application"
          C[MyCustomPlugin]
          C --> |extends| B
      end

      A --> |Provides| D[register method<br/>hooks property<br/>abstract metadata property]
      B --> |Specializes for| E[LDClient type<br/>Flutter SDK context]
      C --> |Implements| F[Custom behavior via override of register<br/>Custom hooks via override of hooks property<br/>Concrete metadata property]

```